Added support for using associated models as columns
- Added the protected function _cointain which accepts a reference of the settings
property as parameter and looks for key settings for defining the associated model:
 bindModel (bool) - weather or no the column is an associated model
 className (string) - will try to get from column name, but can be specified
 displayField (string) - default: name, which column to display
 foreignKey (string) - default: id, in case we use a link
 * createLink (bool) - default: false, weather or not to create a link on the column
 * controller (string) - for link, will try to get form column name, but can be specified
 * action (string) - default: view, can be changed

If createLink is set to true the link will be generated using the controller and action,
and the foreignKey value will be appended at the end.

* Will need to use the datatable/data element as the DataTableResponseView,
just add echo $this->element('datatable/data') on the DataTableResponseView
at the View/{Controller}/datatable/{action_name}.ctp file

- Created the datatable/data.ctp element that dynamically creates the JSON response
based on the DataTable component settings

- Created the datatable/init.ctp element that dynamically creates the settings
from the model/controller and initializes the dataTable..

This elements are ment to be examples, copy and paste on your app and change per need basis.